producer/requirements.txt
# producer/requirements.txt
pika==1.3.2
pandas==2.1.4
numpy==1.26.4  # Actualizado a versión compatible con Python 3.12
python-dotenv==1.0.0
setuptools==70.0.0  # AÑADIDO: incluye distutils